# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Library Management API",
    "description": "API documentation for managing borrowers and books in a library system.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local server"
    }
  ],
  "paths": {
    "/api/borrowers": {
      "post": {
        "summary": "Register a new borrower",
        "description": "Creates a new borrower in the library system.",
        "operationId": "registerBorrower",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Borrower"
              },
              "example": {
                "name": "John Doe",
                "email": "john.doe@example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Borrower registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Borrower"
                }
              }
            }
          }
        },
        "tags": ["Borrowers"]
      }
    },
    "/api/books": {
      "post": {
        "summary": "Register a new book",
        "description": "Creates a new book in the library system.",
        "operationId": "registerBook",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              },
              "example": {
                "isbn": "978-3-16-148410-0",
                "title": "Effective Java",
                "author": "Joshua Bloch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        },
        "tags": ["Books"]
      },
      "get": {
        "summary": "Get all books",
        "description": "Retrieves a list of all books in the library system.",
        "operationId": "getAllBooks",
        "responses": {
          "200": {
            "description": "List of books retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              }
            }
          }
        },
        "tags": ["Books"]
      }
    },
    "/api/books/{bookId}/borrow": {
      "post": {
        "summary": "Borrow a book",
        "description": "Allows a borrower to borrow a specific book by its ID.",
        "operationId": "borrowBook",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "description": "The ID of the book to borrow",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "borrowerId",
            "in": "query",
            "required": true,
            "description": "The ID of the borrower",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book borrowed successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Book borrowed successfully."
                }
              }
            }
          },
          "404": {
            "description": "Book or borrower not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Book not found"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Books"]
      }
    },
    "/api/books/{bookId}/return": {
      "post": {
        "summary": "Return a book",
        "description": "Allows a borrower to return a borrowed book by its ID.",
        "operationId": "returnBook",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "description": "The ID of the book to return",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book returned successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Book returned successfully."
                }
              }
            }
          },
          "404": {
            "description": "Book not found or not borrowed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Book not found"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Books"]
      }
    }
  },
  "components": {
    "schemas": {
      "Borrower": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          }
        }
      },
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "isbn": {
            "type": "string",
            "example": "978-3-16-148410-0"
          },
          "title": {
            "type": "string",
            "example": "Effective Java"
          },
          "author": {
            "type": "string",
            "example": "Joshua Bloch"
          },
          "isBorrowed": {
            "type": "boolean",
            "example": false
          }
        }
      }
    }
  }
}